requirements.txt
streamlit>=1.44.0
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
openai>=1.76.0
jieba>=0.42.1

